refactored according to feedback.
