+++ public/app/core/utils/standardTransformers.ts
@@ -14,7 +14,7 @@ import { seriesToRowsTransformerRegistryItem } from '../components/TransformersU
 import { concatenateTransformRegistryItem } from '../components/TransformersUI/ConcatenateTransformerEditor';
 import { renameByRegexTransformRegistryItem } from '../components/TransformersUI/RenameByRegexTransformer';
 import { histogramTransformRegistryItem } from '../components/TransformersUI/HistogramTransformerEditor';
-import { stretchFrameTransformerRegistryItem } from '../components/TransformersUI/StretchFramesTransformerEditor';
+import { stretchFrameTransformerRegistryItem } from '../components/TransformersUI/stretchFrames/StretchFramesTransformerEditor';
 import { rowsToFieldsTransformRegistryItem } from '../components/TransformersUI/rowsToFields/RowsToFieldsTransformerEditor';
 import { configFromQueryTransformRegistryItem } from '../components/TransformersUI/configFromQuery/ConfigFromQueryTransformerEditor';
